fix(mcp-server): use native Vercel Node server
