+++ apps/mcp-server/package.json
@@ -22,7 +22,7 @@
     "start": "node dist/index.js",
     "dev": "node --import tsx src/index.ts",
     "typecheck": "tsc --noEmit -p tsconfig.check.json",
-    "lint": "biome lint src/ instrumentation.ts",
+    "lint": "biome lint src/ instrumentation.ts server.ts",
     "format": "biome format --write .",
     "test": "vitest run",
     "test:watch": "vitest"
