--- apps/mcp-server/src/server.ts
@@ -1,3 +1,0 @@
-// Compatibility entrypoint for explicit function configurations. Vercel's
-// Node framework packages src/index.ts as the application root.
-export { default } from "./index.js";
