--- apps/mcp-server/vercel.json
@@ -1,9 +1,0 @@
-{
-  "$schema": "https://openapi.vercel.sh/vercel.json",
-  "buildCommand": "bun run build",
-  "functions": {
-    "src/index.ts": {
-      "maxDuration": 60
-    }
-  }
-}
